// src/utils/thirdParty/dayjs/types.ts
/*
 * @Author: czy0729
 * @Date: 2026-09-04 19:22:00
 * @Last Modified by:   czy0729
 * @Last Modified time: 2026-09-04 19:22:00
 */

/** 支持的时间单位 (兼容单复数) */
export type Unit =
  | 'day'
  | 'days'
  | 'week'
  | 'weeks'
  | 'month'
  | 'months'
  | 'hour'
  | 'hours'
  | 'minute'
  | 'minutes'
  | 'second'
  | 'seconds'